Update DHT.cpp to work with DHT22 in proteus.
With this changes we can use this library to work wiith DHT22 in proteus. This sensor don´t read the first 16bits like RH and the other 16bits to temperature, the last 8bits are the checksum. I made this changes in this archive on my laptop and it works.

I´m university student and I don´t speak english so good but I hope this changes are correct.